requirements.txt
streamlit
tensorflow==2.16.1
keras==3.0.0
numpy
Pillow
matplotlib